requirements.txt
langchain-core
langchain-openai
langchain-community
langgraph
numpy
faiss-cpu
pydantic-settings
retry
decorator
fastapi
uvicorn
supabase
psycopg2[binary]